[perfcompare] Remove perfcompare.py's dependency on SciPy and hence on vpython
This allows perfcompare.py to be run using fuchsia-vendored-python, so
that it doesn't require using Infra's vpython (or using a
system-installed SciPy).

This just involves providing our own implementation of
scipy.stats.t.ppf().  This was generated by Gemini.  This
implementation looks reasonable, and its outputs match SciPy to many
decimal places, but I have not checked it in any other significant
way.

Here is the Gemini conversation that produced the code:
https://docs.google.com/document/d/1C0nTwTHUyupaLMp3a9b859QLdg8k2k9XgW2tAzWWRRY/edit

Bug: 496596903
Test: perfcompare_test.py and stats_test.py
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1574160
Commit-Queue: Mark Seaborn <mseaborn@google.com>

+++ src/testing/perfcompare/perfcompare_test.py
@@ -649,10 +649,12 @@ class PerfCompareTest(TempDirTestCase):
 
     def test_factor_range_formatting(self):
         # Construct an interval pair of the same type used in the
-        # software-under-test, checking that the interval is well-formed.
+        # software-under-test, in order to ensure that this test case is
+        # realistic.  (Earlier versions used numpy.float64.)  Also check
+        # that the interval is well-formed.
         def Interval(min_val, max_val):
             assert min_val <= max_val
-            return (numpy.float64(min_val), numpy.float64(max_val))
+            return (float(min_val), float(max_val))
 
         # Check that the values are of the same type as in the
         # software-under-test.
